# doljabi/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "doljabi",
    "description": "doljabi project REST API를 정의한 문서입니다.",
    "contact": {
      "name": "Doljabi Team",
      "email": "doljabi2025@gmail.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/openapi/openapi.json": {
      "get": {
        "tags": [
          "openapi"
        ],
        "summary": "OpenAPI JSON 스펙을 반환하는 엔드포인트",
        "operationId": "get_openapi_spec",
        "responses": {
          "200": {
            "description": "OpenAPI 스펙 반환 성공"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
    "/api/user/games/session/{session_key}": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "get_my_games",
        "parameters": [
          {
            "name": "session_key",
            "in": "path",
            "description": "세션 키",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "내 대국 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListResponse"
                }
              }
            }
          },
          "400": {
            "description": "세션 키가 올바르지 않음"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/api/user/games/{game_id}/session/{session_key}": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "get_my_game_detail",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "description": "게임 고유 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_key",
            "in": "path",
            "description": "세션 키",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게임 상세 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "세션 키가 올바르지 않음"
          },
          "404": {
            "description": "해당 게임이 없거나 권한 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
...
    "/ws/room/{enter_code}/session/{session_key}": {
      "get": {
        "operationId": "enter_room",
        "parameters": [
          {
            "name": "enter_code",
            "in": "path",
            "description": "방 입장 코드",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "session_key",
            "in": "path",
            "description": "세션 키",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket 연결 성공: 방 입장 성공"
          },
          "400": {
            "description": "유효하지 않은 세션 키"
          },
          "404": {
            "description": "존재하지 않는 방"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GameDetail": {
        "type": "object",
        "required": [
          "id",
          "game_type",
          "black_name",
          "white_name",
          "result",
          "played_at",
          "move_count",
          "history"
        ],
        "properties": {
          "black_name": {
            "type": "string"
          },
          "game_type": {
            "type": "string"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MoveRecord"
            }
          },
          "id": {
            "type": "string"
          },
          "move_count": {
            "type": "integer",
            "format": "int32"
          },
          "played_at": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "white_name": {
            "type": "string"
          }
        }
      },
      "GameDetailResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "game": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameDetail"
              }
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "GameListResponse": {
        "type": "object",
        "required": [
          "success",
          "games"
        ],
        "properties": {
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            }
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GameSummary": {
        "type": "object",
        "description": "목록 한 줄에 들어가는 요약.",
        "required": [
          "id",
          "game_type",
          "black_name",
          "white_name",
          "result",
          "played_at",
          "move_count"
        ],
        "properties": {
          "black_name": {
            "type": "string"
          },
          "game_type": {
            "type": "string",
            "description": "\"go\" | \"omok\" (프론트 식별자에 맞춤; \"baduk\"은 \"go\"로 매핑)"
          },
          "id": {
            "type": "string"
          },
          "move_count": {
            "type": "integer",
            "format": "int32"
          },
          "played_at": {
            "type": "string",
            "description": "ISO-8601 (UTC)"
          },
          "result": {
            "type": "string",
            "description": "SGF RE 형식 ('B+R' / 'W+T' / '0' 등)"
          },
          "white_name": {
            "type": "string"
          }
        }
      },
...
      "MoveRecord": {
        "type": "object",
        "description": "수순 한 수. col/row는 0-based, pass면 col<0 && row<0.",
        "required": [
          "col",
          "row",
          "color"
        ],
        "properties": {
          "col": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "description": "\"black\" | \"white\""
          },
          "pass": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "row": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}